Fixing a bug in uninstall.bat
Running uninstall.bat from a junction does not delete the installed folder.
Running it from the original path does not cause any problems.
That is why in d4634bf lines 1611 and 1612 %origintexdir% was written instead of %texdir%.
(These were later renamed: %origintexfpdir% %texfpdir%)
But this causes another bug:
The installer exe takes the path of the previous installation from HKCU\Software\Microsoft\Windows\CurrentVersion\Uninstall\TeXfireplace, so %previousinstalldir%=%origintexfpdir%.
Therefore, the "Remove previous installation" block wants to remove the %origintexfpdir% part from %path%, which is not good in the case of a junction, because then %path% contains %texfpdir%.
The solution is to write the %texfpdir% path back to HKCU\Software\Microsoft\Windows\CurrentVersion\Uninstall\TeXfireplace and fix uninstall.bat to a version that can also delete the installed folder via a junction.